Simplify OfferPresenterNew
